requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
einops>=0.6.0

pyarrow>=12.0.0  # for parquet files
Pillow>=9.0.0  # for image processing
requests>=2.28.0
tqdm>=4.65.0
colt5_attention>=0.1.0

transformers>=4.30.0  # for ViT and BERT models
torchvision>=0.15.0  # for image transforms

tensorboard>=2.12.0  # for TensorBoard writer

# Optional dependencies - comment out if causing installation issues
# flash-attn  # for flash attention (optional)